Trim code back to the essentials for now.
I'm in the process of tweaking how qualities work to require,
hopefully, less verbose coding on down the line. What I'm commiting
here, the engine works, but the MUDSocket and test MUD don't:

The next step is changing set-thing-quality! to accept an optional
boolean fourth argument, default false, but if true, it adds the
quality to the thing. (And if false, and the thing doesn't have a
quality, then set-thing-quality! itself returns false.