Adding look, move, features
So this is the first real test of the new way of handling working with
qualities, where the default handling procedures (thing-quality,
set-thing-quality!, add-element-to-thing-quality!, and so on) look at
the things being worked with, and their universes, for more specific
procedures.

In this use, a MUD can (add-animation-procedures-to-universe!
their-universe) which adds the add-element-to-thing-contents!
procedure to their-universe.

Then, when (add-element-to-thing-quality! first-thing second-thing
'contents) is called, it passes on to (add-element-to-thing-contents!
first-thing second-thing), which does fancy stuff like checking the
first-thing's container quality and updating that container thing, if
it has one.

This is the plan for handling lots of stuff: pass it on to a more
specific procedure, closer to the action.

Next up is trying to make "move" work, which'll probably mean writing
other specific procedures like that.